// ui/src/app/chart-shared/components/charts/progress-bar/model/progress-bar-widget.model.ts
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

import {
    DataExplorerDataConfig,
    DataExplorerField,
    DataExplorerWidgetModel,
} from '@streampipes/platform-services';
import {
    DataExplorerVisConfig,
    WidgetNumberAppearanceConfig,
} from '../../../../models/dataview-dashboard.model';

export type ProgressBarTargetSource = 'fixed' | 'field';
export type ProgressBarDisplayMode = 'percent' | 'value' | 'percent-and-value';

export interface ProgressBarVisConfig extends DataExplorerVisConfig {
    currentValueField?: DataExplorerField;
    targetSource?: ProgressBarTargetSource;
    targetField?: DataExplorerField;
    targetValue?: number;
    invertProgress?: boolean;
    clampProgress?: boolean;
    displayMode?: ProgressBarDisplayMode;
    showLabel?: boolean;
    title?: string;
    description?: string;
}

export interface ProgressBarAppearanceConfig extends WidgetNumberAppearanceConfig {
    progressColor?: string;
    trackColor?: string;
}

export interface ProgressBarWidgetModel extends DataExplorerWidgetModel {
    dataConfig: DataExplorerDataConfig;
    visualizationConfig: ProgressBarVisConfig;
    baseAppearanceConfig: ProgressBarAppearanceConfig;
}